requirements.txt
streamlit==1.26.0
tensorflow==2.12.0
numpy==1.24.3
Pillow==9.4.0
